trusty: Add VtsHalGatekeeperTargetTest
Commit d0149e8e9a129b211e3518001e735d429287d99b in system/core changed
Gatekeeper from HIDL to AIDL. This requires a new vts test to the built
(and run the old test succeeds without testing anything). Add that test
to this build. The test scripts will need to be updated in a separate
change once a prebuilt image with this change is ready.

Bug: 268342724
Test: VtsHalGatekeeperTargetTest
Change-Id: I847e2dca212b5f8122587e7b1aca2df7c6634a50