Set ro.product.first_api_level on qemu
ro.product.first_api_level is used to determine the API level that a
given device launched with. Certain provisioning tests depend on this
directly to understand what the vendor API level was when the device was
first provisioned, as opposed to the more common need of understanding
the vendor API level that a device currently conforms to.

Test: adb shell getprop ro.product.first_api_level
Test: # Note this suite still doesn't pass but tests aren't
      # incorrectly skipped
      atest VtsHalRemotelyProvisionedComponentTargetTest
Bug: 418221547
Change-Id: I036c488100bbb29294cfadb2470ab6230c531064