memshare_vm: app: End-to-End sample code
Adding sample code for an end to end example that
allows a VM to use buffers created on Android. It
includes a curtom system VM launcher, a sample
Android application that creates a buffer and sends
it to the VM and a set of AIDLs that describe the
interface between the components.

Bug: 454503409
Test: m app.test
Change-Id: I10ecc1fa315a682726ebca57367bd863cbe7bed6