Add ISharedSecret defauld and shared_secret apex
Adding apex packages for HwCrypto. Currently the packages only have
vintf fragments and .rc file that set up a property indicating
if the HwCrypto HAL service should register the ISharedSecret
HAL.
The reason for this is that currently if we add the service on
the apex the system complains about a dependency on libbinder (the
service is a bridge between cpp and ndk backends).

Bug: 465731563
Test: atest VtsAidlSharedSecretTargetTest
Change-Id: I4cc522862ea9fec416a4a2906348968ff74c4386